requirements.txt
hydra-core==1.1.1
numpy==1.22.2
omegaconf==2.1.1
pelutils==0.6.9
torch==1.10.2
torchvision==0.11.3
wandb==0.12.10
pelutils==0.6.9
psutil==5.9.0
requests==2.27.1

# Server stuff
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
